Move Pre around, use Highlight component
