+++ src/constants/code-snippets.js
@@ -1,0 +1,1 @@
+export const CODE_SNIPPET_CLASSNAME = 'frontpage-code-snippets';
